Face touch example

+++ examples/facetouch/src/App.tsx
@@ -18,10 +18,11 @@ import Slider from '@material-ui/core/Slider';
 import Grid from '@material-ui/core/Grid';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Checkbox from '@material-ui/core/Checkbox';
+import FormControl from '@material-ui/core/FormControl';
+import InputLabel from '@material-ui/core/InputLabel';
+import MenuItem from '@material-ui/core/MenuItem';
+import Select from '@material-ui/core/Select';
 
-const numTrainSamples = 64;
-const numValidationSamples = 0;
-const numSamples = numTrainSamples + numValidationSamples;
 //const featureDim = 64;
 const featureDim = 1280;
 
@@ -36,6 +37,9 @@ interface AppState {
   numIterations: number;
   sound: boolean;
   notification: boolean;
+  model: string;
+  numTrainSamples: number;
+  numValidationSamples: number;
 }
 
 function wait(t: number) {
@@ -54,24 +58,42 @@ async function countDown(n: number, t: number, cb: (n: number) => void) {
   cb(0);
 }
 
+const models = [
+  {
+    size: 1.0,
+    name: 'mobilenet100'
+  },
+  {
+    size: 0.5,
+    name: 'mobilenet050'
+  },
+  {
+    size: 0.35,
+    name: 'mobilenet035'
+  },
+  {
+    size: 0.25,
+    name: 'mobilenet025',
+    outputs: ['473'],
+    prune: ['474']
+  }
+]
+
 class App extends React.Component<{}, AppState> {
   private model?: tjs.onnx.model.OnnxModel = undefined;
-  private classifier?: tjs.model.Module;
-  private modelCompiled = false;
+  private classifier?: tjs.model.basic.Linear;
 
   private mean?: tjs.Tensor;
   private varSqrt?: tjs.Tensor;
 
-  private meanMobilenet = new tjs.tensor.gpu.GPUTensor(new Float32Array([0.485, 0.456, 0.406]), [3,1,1], 16);
-  private stdMobilenet = new tjs.tensor.gpu.GPUTensor(new Float32Array([0.229, 0.224, 0.225]), [3,1,1], 16);
+  private meanMobilenet = new tjs.tensor.gpu.GPUTensor(new Float32Array([0.485, 0.456, 0.406]), [3,1,1], 32);
+  private stdMobilenet = new tjs.tensor.gpu.GPUTensor(new Float32Array([0.229, 0.224, 0.225]), [3,1,1], 32);
 
-  private data = new tjs.tensor.gpu.GPUTensor(new Float32Array(numSamples*featureDim).fill(0), [numSamples, featureDim], 32);
+  private data?: tjs.tensor.gpu.GPUTensor;
 
   private trainingResultsCollected = 0;
 
   private videoTensor?: tjs.tensor.gpu.GPUTensor;
-
-  private start = Date.now().valueOf();
 
   private audio?: HTMLAudioElement;
 
@@ -86,19 +108,73 @@ class App extends React.Component<{}, AppState> {
       threshold: 0.5,
       sound: true,
       notification: false,
+      model: 'mobilenet050',
+      numTrainSamples: 64,
+      numValidationSamples: 0
     };
   }
 
   componentWillMount() {
-    loadModel('mobilenet050').then(x => {
-      if (x !== undefined) {
-        this.model = x;
-
-        this.getVideo();
-      }
-    });
+    this.setModel('mobilenet050');
 
     this.audio = new Audio('alerts/beep.wav');
+  }
+
+  componentDidMount() {
+    this.getVideo().then(x => {
+      this.checkStorage();
+    });
+  }
+
+  async checkStorage() {
+    const model = localStorage.getItem('model');
+    const mean = localStorage.getItem('mean');
+    const varSqrt = localStorage.getItem('varSqrt');
+    const classifierValues = localStorage.getItem('classifierValues');
+
+    if (model !== null && mean !== null && varSqrt !== null && classifierValues !== null) {
+      const parsedClassifier = JSON.parse(classifierValues);
+
+      this.classifier = new Linear(featureDim, 1, true);
+      this.classifier.weights = Variable.create([featureDim, 1], parsedClassifier[0], 'GPU');
+      this.classifier.bias = Variable.create([1], parsedClassifier[1], 'GPU');
+
+      this.mean = new tjs.tensor.gpu.GPUTensor(new Float32Array(JSON.parse(mean)), [featureDim], 32);
+      this.varSqrt = new tjs.tensor.gpu.GPUTensor(new Float32Array(JSON.parse(varSqrt)), [featureDim], 32);
+
+      await this.setModel(model);
+      this.setState({
+        stage: 'monitoring',
+        prediction: 0,
+      });
+
+      await wait(1000);
+
+      const reshaped = this.prepareVideo() as tjs.Tensor;
+
+      this.model?.forward([reshaped]).then(results => {
+        reshaped.delete();
+        this.process(results);
+      });
+    }
+  }
+
+  async setModel(name: string) {
+    const model = models.find(x => x.name === name);
+
+    this.setState({
+      model: name,
+      stage: 'start'
+    });
+
+    this.model = await loadModel(name);
+
+    if (this.model !== undefined) {
+      this.model.outputs = model?.outputs || ['472'];
+      this.model.prune(model?.prune || ["473"]);
+    }
+
+    this.model?.optimize();
   }
 
   async getVideo() {
@@ -116,51 +192,43 @@ class App extends React.Component<{}, AppState> {
   }
 
   prepareVideo() {
-    if (this.videoTensor !== undefined) {
-      const video: HTMLVideoElement = (document.querySelector("#videoElement") as any);
+  const video: HTMLVideoElement = (document.querySelector("#videoElement") as any);
 
-      this.videoTensor = tjs.tensor.gpu.GPUTensor.fromData(video, 32);
+    this.videoTensor = tjs.tensor.gpu.GPUTensor.fromData(video, 32);
 
-      let [height, width] = this.videoTensor.shape.slice(0,2);
+    let [height, width] = this.videoTensor.shape.slice(0,2);
 
-      const sliced = this.videoTensor.slice([0], [3], [2]);
+    const sliced = this.videoTensor.slice([0], [3], [2]);
 
-      this.videoTensor.delete();
+    this.videoTensor.delete();
 
-      const transposed = sliced.transpose([2, 0, 1]);
-      sliced.delete();
+    const transposed = sliced.transpose([2, 0, 1]);
+    sliced.delete();
 
-      const scaled = transposed.subtract(this.meanMobilenet);
-      transposed.delete();
-      const normalized = scaled.divide(this.stdMobilenet);
-      scaled.delete();
+    const scaled = transposed.subtract(this.meanMobilenet);
+    transposed.delete();
+    const normalized = scaled.divide(this.stdMobilenet);
+    scaled.delete();
 
-      const reshaped = normalized.reshape([1,3,height,width], false);
-      return reshaped;
-    }
+    const reshaped = normalized.reshape([1,3,height,width], false);
+    return reshaped;
   }
 
-  async compileModel() {
+  async prepareTraining() {
     const video: HTMLVideoElement = (document.querySelector("#videoElement") as any);
     this.videoTensor = tjs.tensor.gpu.GPUTensor.fromData(video, 32);
-    if (this.videoTensor !== undefined && !this.modelCompiled) {
+    if (this.videoTensor !== undefined) {
       this.setState({
         ...this.state,
         stage: 'compiling'
       });
 
-      this.modelCompiled = true;
+      const numSamples = this.state.numTrainSamples + this.state.numValidationSamples;
 
       const reshaped = this.prepareVideo() as tjs.Tensor;
 
-      if (this.model !== undefined) {
-        this.model.outputs = ["472"];
-        this.model.prune(["473"]);
-      }
+      this.data = new tjs.tensor.gpu.GPUTensor(new Float32Array(numSamples*featureDim).fill(0), [numSamples, featureDim], 32);
 
-      this.model?.optimize();
-
-      console.log('Compiled');
       console.log('Doing forward pass');
 
       await countDown(3, 1000, (n: number) => {
@@ -191,9 +259,13 @@ class App extends React.Component<{}, AppState> {
     const processed = this.processResult(tensors);
 
     const oldResults = this.data;
+    //@ts-ignore
     this.data = this.data.setValues(processed, [this.trainingResultsCollected, 0]) as tjs.tensor.gpu.GPUTensor;
     processed.delete();
+    //@ts-ignore
     oldResults.delete();
+
+    const numSamples = this.state.numTrainSamples + this.state.numValidationSamples;
 
     this.trainingResultsCollected++;
 
@@ -275,12 +347,18 @@ class App extends React.Component<{}, AppState> {
   }
 
   async prepareData() {
-    const trainData1 = this.data.slice([0],[numTrainSamples/2],[0]);
-    const trainData2 = this.data.slice([numSamples/2],[numSamples/2 + numTrainSamples/2],[0]);
+    const numSamples = this.state.numTrainSamples + this.state.numValidationSamples;
+
+    //@ts-ignore
+    const trainData1 = this.data.slice([0],[this.state.numTrainSamples/2],[0]);
+    //@ts-ignore
+    const trainData2 = this.data.slice([numSamples/2],[numSamples/2 + this.state.numTrainSamples/2],[0]);
     let trainData = trainData1.concat(trainData2, 0);
 
-    const valData1 = this.data.slice([numTrainSamples/2],[numSamples/2],[0]);
-    const valData2 = this.data.slice([numSamples/2 + numTrainSamples/2],[numSamples],[0]);
+    //@ts-ignore
+    const valData1 = this.data.slice([this.state.numTrainSamples/2],[numSamples/2],[0]);
+    //@ts-ignore
+    const valData2 = this.data.slice([numSamples/2 + this.state.numTrainSamples/2],[numSamples],[0]);
     let valData = valData1.concat(valData2, 0);
 
     trainData = await this.normalizeResults(trainData);
@@ -292,11 +370,11 @@ class App extends React.Component<{}, AppState> {
     valData = normalizedValData;
 
     const trainX = new Variable(trainData, {noGrad: true});
-    const trainYs = [...new Array(numTrainSamples/2).fill(0),...new Array(numTrainSamples/2).fill(1)]
-    const trainY = Variable.create([numTrainSamples, 1], new Float32Array(trainYs), 'GPU', {noGrad: true});
+    const trainYs = [...new Array(this.state.numTrainSamples/2).fill(0),...new Array(this.state.numTrainSamples/2).fill(1)]
+    const trainY = Variable.create([this.state.numTrainSamples, 1], new Float32Array(trainYs), 'GPU', {noGrad: true});
 
     const valX = new Variable(valData, {noGrad: true});
-    const valYs = [...new Array(numValidationSamples/2).fill(0),...new Array(numValidationSamples/2).fill(1)]
+    const valYs = [...new Array(this.state.numValidationSamples/2).fill(0),...new Array(this.state.numValidationSamples/2).fill(1)]
 
     return {trainX, trainYs, trainY, valX, valYs};
   }
@@ -355,12 +433,32 @@ class App extends React.Component<{}, AppState> {
 
     await wait(1000);
 
+    this.saveModelSettings();
+
     const reshaped = this.prepareVideo() as tjs.Tensor;
 
     this.model?.forward([reshaped]).then(results => {
       reshaped.delete();
       this.process(results);
     });
+  }
+
+  async saveModelSettings() {
+    const params = this.classifier?.getParameters() as Variable[];
+    const paramValues = [];
+    for (let param of params) {
+      paramValues.push(Array.from(await param.getValues()));
+    }
+
+    localStorage.setItem('classifierValues', JSON.stringify(paramValues));
+
+    const meanValues = Array.from(await this.mean?.getValues() as any);
+    localStorage.setItem('mean', JSON.stringify(meanValues));
+
+    const stdValues = Array.from(await this.varSqrt?.getValues() as any);
+    localStorage.setItem('varSqrt', JSON.stringify(stdValues));
+
+    localStorage.setItem('model', this.state.model);
   }
 
   toggleNotification(value: boolean) {
@@ -421,14 +519,16 @@ class App extends React.Component<{}, AppState> {
       this.numIts = 0;
     }
 
-    await wait(1000);
+    if (this.state.stage === 'monitoring') {
+      await wait(1000);
 
-    const reshaped = this.prepareVideo() as tjs.Tensor;
+      const reshaped = this.prepareVideo() as tjs.Tensor;
 
-    this.model?.forward([reshaped]).then(results => {
-      reshaped.delete();
-      this.process(results);
-    });
+      this.model?.forward([reshaped]).then(results => {
+        reshaped.delete();
+        this.process(results);
+      });
+    }
   }
 
   getCorrect(predVals: number[], ys: number[]) {
@@ -463,7 +563,16 @@ class App extends React.Component<{}, AppState> {
 
   renderSettings() {
     return (
-      <Accordion className="settings">
+      <div className="settings">
+        {this.renderGeneralSettings()}
+        {this.renderModelSettings()}
+      </div>
+    );
+  }
+
+  renderGeneralSettings() {
+    return (
+      <Accordion>
         <AccordionSummary
           aria-controls="panel1a-content"
           id="panel1a-header"
@@ -536,9 +645,86 @@ class App extends React.Component<{}, AppState> {
     );
   }
 
+  renderModelSettings() {
+    const marks = [
+      {
+        value: 16,
+        label: '16',
+      },
+      {
+        value: 32,
+        label: '32',
+      },
+      {
+        value: 64,
+        label: '64',
+      },
+      {
+        value: 128,
+        label: '128',
+      },
+      {
+        value: 256,
+        label: '256',
+      },
+    ];
+
+    return (
+      <Accordion>
+        <AccordionSummary
+          aria-controls="panel1a-content"
+          id="panel1a-header"
+        >
+          <Typography variant="h5">Model</Typography>
+        </AccordionSummary>
+        <AccordionDetails>
+          <Grid container spacing={1}>
+            <Grid item xs={12}>
+              <Typography id="slider-threshold-label" gutterBottom>
+                Model Size
+              </Typography><div></div>
+            </Grid>
+            <Grid item xs={12}>
+              <FormControl>
+                <InputLabel id="label-model-size">Age</InputLabel>
+                <Select
+                  labelId="label-model-size"
+                  id="model-size"
+                  value={this.state.model}
+                  onChange={(event) => this.setModel(event.target.value as any)}
+                >
+                  {models.map(x => (
+                    <MenuItem value={x.name}>{x.size}</MenuItem>
+                  ))}
+                </Select>
+              </FormControl>
+            </Grid>
+            <Grid item xs={12}>
+              <Typography id="slider-train-samples-label" gutterBottom>
+                Number of training samples
+              </Typography><div></div>
+            </Grid>
+            <Grid item xs={12}>
+              <Slider
+                aria-labelledby="slider-train-samples-label"
+                value={this.state.numTrainSamples}
+                step={null}
+                valueLabelDisplay="auto"
+                marks={marks}
+                min={16}
+                max={256}
+                onChange={(ev, v) => this.setState({numTrainSamples: v as number})}
+              />
+            </Grid>
+          </Grid>
+        </AccordionDetails>
+      </Accordion>
+    );
+  }
+
   renderState() {
     if (this.state.stage === 'start') {
-      return (<Button variant="contained" color="primary" onClick={() => this.compileModel()}>Start</Button>);
+      return (<Button variant="contained" color="primary" onClick={() => this.prepareTraining()}>Start</Button>);
     } else if (this.state.stage === 'compiling') {
       return (<> Dont touch your face until the bar is full. Starting in {this.state.countDown} seconds. </>);
     } else if (this.state.stage === 'trainData-1' || this.state.stage === 'trainData-2') {
